requirements.txt
Django==5.2.4
asgiref==3.9.1
django-environ==0.12.0
pillow==11.3.0
psycopg2-binary==2.9.6
boto3==1.34.0
django-storages==1.14.0
python-dotenv==1.0.0
gunicorn==20.1.0
sqlparse==0.5.3
tzdata==2025.2
